Moved the cube plotting code to main.py. Added extra comments
